ENH: now correctly scales the location of the polygon points drawn on all 3 slice nodes.
Seed node is added to the scene when the user clicks the button (from the widget), then when the user clicks the spot to place the seed the position is set and this causes a modified event. The DM then creates the representations for all 3 views (since there is a callback for each slice node).